requirements.txt
openai==1.77.0
pandas==2.2.3
Pillow==11.2.1
praw==7.8.1
pytesseract==0.3.13
python-dotenv==1.1.0
streamlit==1.42.0
